performance-class {ROCR}	R Documentation

Class "performance"

Description

Object to capture the result of a performance evaluation, optionally collecting evaluations from several cross-validation or bootstrapping runs.

Details

A performance object can capture information from four different evaluation scenarios:

The behaviour of a cutoff-dependent performance measure across the range of all cutoffs (e.g. performance( predObj, 'acc' ) ). Here, x.values contains the cutoffs, y.values the corresponding values of the performance measure, and alpha.values is empty.

The trade-off between two performance measures across the range of all cutoffs (e.g. performance( predObj, 'tpr', 'fpr' ) ). In this case, the cutoffs are stored in alpha.values, while x.values and y.values contain the corresponding values of the two performance measures.

A performance measure that comes along with an obligatory second axis (e.g. performance( predObj, 'ecost' ) ). Here, the measure values are stored in y.values, while the corresponding values of the obligatory axis are stored in x.values, and alpha.values is empty.

A performance measure whose value is just a scalar (e.g. performance( predObj, 'auc' ) ). The value is then stored in y.values, while x.values and alpha.values are empty.

Objects from the Class

Objects can be created by using the performance function.

Slots

x.name:

Performance measure used for the x axis.

y.name:

Performance measure used for the y axis.

alpha.name:

Name of the unit that is used to create the parametrized curve. Currently, curves can only be parametrized by cutoff, so alpha.name is either none or cutoff.

x.values:

A list in which each entry contains the x values of the curve of this particular cross-validation run. x.values[[i]], y.values[[i]], and alpha.values[[i]] correspond to each other.

y.values:

A list in which each entry contains the y values of the curve of this particular cross-validation run.

alpha.values:

A list in which each entry contains the cutoff values of the curve of this particular cross-validation run.

Author(s)

Tobias Sing tobias.sing@mpi-sb.mpg.de, Oliver Sander osander@mpi-sb.mpg.de

References

A detailed list of references can be found on the ROCR homepage at http://rocr.bioinf.mpi-sb.mpg.de.

See Also

prediction, performance, prediction-class, plot.performance

[Package ROCR version 1.0-2 Index]